mcp-pytest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-pytest"
version = "0.1.0"
description = "A pytest plugin for testing MCP (Model Context Protocol) servers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["pytest", "mcp", "testing", "model-context-protocol", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "mcp>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-html>=4.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
    "pytest-cov",
]

[project.entry-points.pytest11]
mcp_pytest = "mcp_pytest.plugin"

[project.urls]
Homepage = "https://github.com/yourname/mcp-pytest"
Documentation = "https://github.com/yourname/mcp-pytest#readme"
Repository = "https://github.com/yourname/mcp-pytest"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_pytest"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
